query types and constants
